data-stream-processor/pyproject.toml
[tool.poetry]
name = "data-stream-processor"
version = "0.1.0"
description = ""
authors = ["RedwanulKarim <karim.redwanul.de@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
grpcio-tools = "^1.59.3"
grpcio = "^1.59.3"
loguru = "^0.7.2"
black = "^23.11.0"
asyncio = "^3.4.3"
apache-flink = "^1.18.0"
protobuf = "^4.23.4"
google-cloud-bigquery = "^3.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
